chore(suite): remove unused layoutcomponents and rename mobile menu components
