+++ packages/suite/src/components/suite/Preloader/SuiteLayout/MobileMenu/MobileActionItem.tsx
@@ -1,0 +1,158 @@
+import { useMemo, ReactNode, HTMLAttributes } from 'react';
+import styled, { useTheme } from 'styled-components';
+import { Icon, IconProps, variables, Spinner } from '@trezor/components';
+import { FADE_IN } from '@trezor/components/src/config/animations';
+
+const MobileWrapper = styled.div<Pick<MobileActionItemProps, 'isActive'>>`
+    display: flex;
+    position: relative;
+    cursor: pointer;
+    align-items: center;
+
+    & + & {
+        border-top: 1px solid ${({ theme }) => theme.STROKE_GREY};
+    }
+`;
+
+const MobileIconWrapper = styled.div`
+    display: flex;
+    position: relative;
+    cursor: pointer;
+    align-items: center;
+    margin-right: 16px;
+`;
+
+const Label = styled.span`
+    padding: 16px 8px;
+    color: ${({ theme }) => theme.TYPE_LIGHT_GREY};
+    font-weight: ${variables.FONT_WEIGHT.MEDIUM};
+    font-size: ${variables.FONT_SIZE.NORMAL};
+`;
+
+const AlertDotWrapper = styled.div`
+    position: absolute;
+    top: 9px;
+    right: 10px;
+    width: 9px;
+    height: 9px;
+    border-radius: 50%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    background: ${({ theme }) => theme.BG_WHITE};
+    animation: ${FADE_IN} 0.2s ease-out;
+
+    ${variables.SCREEN_QUERY.BELOW_TABLET} {
+        top: 0;
+        right: 0;
+    }
+`;
+
+const AlertDot = styled.div`
+    position: relative;
+    width: 5px;
+    height: 5px;
+    border-radius: 50%;
+    background: ${({ theme }) => theme.TYPE_ORANGE};
+`;
+
+const Indicator = styled.div`
+    background: ${({ theme }) => theme.BG_WHITE};
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    height: 10px;
+    width: 10px;
+    border-radius: 50%;
+    position: absolute;
+    top: 9px;
+    right: 10px;
+
+    ${variables.SCREEN_QUERY.BELOW_TABLET} {
+        top: 0;
+        right: 0;
+    }
+
+    svg {
+        animation: ${FADE_IN} 0.2s ease-out;
+    }
+`;
+
+export type IndicatorStatus = 'check' | 'alert' | 'loading';
+
+interface CommonProps extends Pick<HTMLAttributes<HTMLDivElement>, 'onClick'> {
+    label: ReactNode;
+    isActive?: boolean;
+    indicator?: IndicatorStatus;
+    'data-test'?: string;
+}
+
+interface CustomIconComponentProps extends CommonProps {
+    iconComponent: ReactNode;
+    icon?: never;
+}
+interface IconComponentProps extends CommonProps {
+    icon: IconProps['icon'];
+    iconComponent?: never;
+}
+
+type MobileActionItemProps = CustomIconComponentProps | IconComponentProps;
+
+// Reason to use forwardRef: We want the user to be able to close Notifications dropdown by clicking somewhere else.
+// In order to achieve that behavior, we need to pass reference to MobileActionItem
+export const MobileActionItem = ({
+    icon,
+    iconComponent,
+    indicator,
+    isActive,
+    label,
+    onClick,
+    'data-test': dataTest,
+}: MobileActionItemProps) => {
+    const theme = useTheme();
+
+    const IconComponent = useMemo(
+        () =>
+            icon ? (
+                <Icon
+                    color={isActive ? theme.TYPE_DARK_GREY : theme.TYPE_LIGHT_GREY}
+                    size={24}
+                    icon={icon}
+                />
+            ) : (
+                iconComponent
+            ),
+        [icon, iconComponent, theme, isActive],
+    );
+
+    const Content = useMemo(
+        () => (
+            <>
+                {IconComponent}
+                {indicator === 'alert' && (
+                    <AlertDotWrapper>
+                        <AlertDot />
+                    </AlertDotWrapper>
+                )}
+                {indicator === 'loading' && (
+                    <Indicator>
+                        <Spinner size={6} />
+                    </Indicator>
+                )}
+                {indicator === 'check' && (
+                    <Indicator>
+                        <Icon icon="CHECK" size={10} color={theme.TYPE_GREEN} />
+                    </Indicator>
+                )}
+            </>
+        ),
+        [indicator, IconComponent, theme],
+    );
+
+    return (
+        <MobileWrapper data-test={dataTest} onClick={onClick}>
+            <MobileIconWrapper>{Content}</MobileIconWrapper>
+            <Label>{label}</Label>
+        </MobileWrapper>
+    );
+};
