+++ packages/suite/src/components/suite/WelcomeLayout/WelcomeLayout.tsx
@@ -15,7 +15,7 @@ import { TREZOR_URL, SUITE_URL } from '@trezor/urls';
 import { resolveStaticPath } from '@suite-common/suite-utils';
 import { GuideButton, GuideRouter } from 'src/components/guide';
 import { useGuide } from 'src/hooks/guide';
-import { NavSettings } from 'src/components/suite/Preloader/SuiteLayout/NavigationBar/NavigationActions/NavSettings';
+import { NavSettings } from 'src/components/suite/WelcomeLayout/NavSettings';
 
 const Wrapper = styled.div`
     display: flex;
